Little style fix
